# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/healthz": {
            "get": {
                "description": "Health.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitoring"
                ],
                "summary": "Health.",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/readyz": {
            "get": {
                "description": "Ready.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitoring"
                ],
                "summary": "Ready.",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/map/{map_name}/generate": {
            "get": {
                "description": "Generate json file from map file.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Map"
                ],
                "summary": "Generate json file from map file.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Map name",
                        "name": "map_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/map/{map_name}/synchronize": {
            "post": {
                "description": "Synchronize GCS file to an HAProxy map file.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Map"
                ],
                "summary": "Synchronize GCS file to an HAProxy map file.",
                "parameters": [
                    {
                        "description": "Data of the synchronisation endpoint",
                        "name": "_",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SynchronizeRequestBody"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Map name",
                        "name": "map_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.SynchronizeRequestBody": {
            "type": "object",
            "required": [
                "bucket_file_name",
                "bucket_name"
            ],
            "properties": {
                "bucket_file_name": {
                    "type": "string"
                },
                "bucket_name": {
                    "type": "string"
                }
            }
        }
    }
}